+++ src/Tools/Routes.ps1
@@ -56,9 +56,11 @@ function Get-PodeRoute
 
     # otherwise, attempt to match on regex parameters
     else {
-        $valid = ($method.Keys | Where-Object {
-            $Route -imatch "^$($_)$"
-        } | Select-Object -First 1)
+        $valid = @(foreach ($key in $method.Keys) {
+            if ($Route -imatch "^$($key)$") {
+                $key
+            }
+        })[0]
 
         if ($null -eq $valid) {
             return $null
@@ -147,7 +149,7 @@ function Get-PodeStaticRoutePath
     }
 
     # else, use the public static directory (but only if path is a file, and a public dir is present)
-    elseif ((Test-PodePathIsFile $Route) -and !(Test-Empty $PodeContext.Server.InbuiltDrives['public'])) {
+    elseif ((Test-PodePathIsFile $Route) -and ![string]::IsNullOrWhiteSpace($PodeContext.Server.InbuiltDrives['public'])) {
         $path = (Join-Path $PodeContext.Server.InbuiltDrives['public'] $Route)
     }
 
@@ -174,13 +176,21 @@ function Get-PodeRouteByUrl
         $Endpoint
     )
 
-    return (@($Routes) |
-        Where-Object {
-            ($_.Protocol -ieq $Protocol -or [string]::IsNullOrEmpty($_.Protocol)) -and
-            ([string]::IsNullOrEmpty($_.Endpoint) -or $Endpoint -ilike $_.Endpoint)
-        } |
-        Sort-Object -Property { $_.Protocol }, { $_.Endpoint } -Descending |
-        Select-Object -First 1)
+    # get the value routes
+    $rs = @(foreach ($route in $Routes) {
+        if (
+            (($route.Protocol -ieq $Protocol) -or [string]::IsNullOrWhiteSpace($route.Protocol)) -and
+            ([string]::IsNullOrWhiteSpace($route.Endpoint) -or ($Endpoint -ilike $route.Endpoint))
+        ) {
+            $route
+        }
+    })
+
+    if ($null -eq $rs[0]) {
+        return $null
+    }
+
+    return @($rs | Sort-Object -Property { $_.Protocol }, { $_.Endpoint } -Descending)[0]
 }
 
 function Route
@@ -238,6 +248,11 @@ function Route
         [string]
         $ErrorType,
 
+        [Parameter()]
+        [Alias('fp')]
+        [string]
+        $FilePath,
+
         [switch]
         [Alias('rm')]
         $Remove,
@@ -293,7 +308,7 @@ function Route
 
         # add the route
         Add-PodeRoute -HttpMethod $HttpMethod -Route $Route -Middleware $Middleware -ScriptBlock $ScriptBlock `
-            -Protocol $Protocol -Endpoint $Endpoint -ContentType $ContentType -ErrorType $ErrorType
+            -Protocol $Protocol -Endpoint $Endpoint -ContentType $ContentType -ErrorType $ErrorType -FilePath $FilePath
     }
 }
 
@@ -380,25 +395,44 @@ function Add-PodeRoute
 
         [Parameter()]
         [string]
-        $ErrorType
+        $ErrorType,
+
+        [Parameter()]
+        [string]
+        $FilePath
     )
 
-    # if middleware and scriptblock are null, error
-    if ((Test-Empty $Middleware) -and (Test-Empty $ScriptBlock)) {
-        throw "[$($HttpMethod)] $($Route) has no logic defined"
+    # if middleware, scriptblock and file path are all null/empty, error
+    if ((Test-Empty $Middleware) -and (Test-Empty $ScriptBlock) -and (Test-Empty $FilePath)) {
+        throw "[$($HttpMethod)] $($Route) has no scriptblock defined"
     }
 
-    # ensure middleware is either a scriptblock, or a valid hashtable
+    # if both a scriptblock and a file path have been supplied, error
+    if (!(Test-Empty $ScriptBlock) -and !(Test-Empty $FilePath)) {
+        throw "[$($HttpMethod)] $($Route) has both a ScriptBlock and a FilePath defined"
+    }
+
+    # if we have a file path supplied, load that path as a scriptblock
+    if (Test-PodePath -Path $FilePath -NoStatus) {
+        # if the path is a wildcard or directory, error
+        if (!(Test-PodePathIsFile -Path $FilePath -FailOnWildcard)) {
+            throw "[$($HttpMethod)] $($Route) cannot have a wildcard or directory FilePath: $($FilePath)"
+        }
+
+        $ScriptBlock = [scriptblock](load $FilePath)
+    }
+
+    # ensure supplied middlewares are either a scriptblock, or a valid hashtable
     if (!(Test-Empty $Middleware)) {
         @($Middleware) | ForEach-Object {
             $_type = (Get-PodeType $_).Name
 
-            # is the type valid
+            # check middleware is a type valid
             if ($_type -ine 'scriptblock' -and $_type -ine 'hashtable') {
                 throw "A middleware supplied for the '[$($HttpMethod)] $($Route)' route is of an invalid type. Expected either ScriptBlock or Hashtable, but got: $($_type)"
             }
 
-            # is the hashtable valid
+            # if middleware is hashtable, ensure the keys are valid (logic is a scriptblock)
             if ($_type -ieq 'hashtable') {
                 if ($null -eq $_.Logic) {
                     throw "A Hashtable middleware supplied for the '[$($HttpMethod)] $($Route)' route has no Logic defined"
@@ -412,7 +446,7 @@ function Add-PodeRoute
         }
     }
 
-    # if middleware set, but not scriptblock, set middle and script
+    # if middleware is set, but there is no scriptblock, set the middleware as the scriptblock
     if (!(Test-Empty $Middleware) -and ($null -eq $ScriptBlock)) {
         # if multiple middleware, error
         if ((Get-PodeType $Middleware).BaseName -ieq 'array' -and (Get-PodeCount $Middleware) -ne 1) {